Postcode is optional in some countries
