+++ support-frontend/assets/pages/digital-subscription-checkout/digitalSubscriptionCheckoutReducer.js
@@ -15,20 +15,19 @@ import {
   stateProvinceFromString,
 } from 'helpers/internationalisation/country';
 import { GBPCountries } from 'helpers/internationalisation/countryGroup';
-import { formError, type FormError, nonEmptyString, notNull, validate } from 'helpers/subscriptionsForms/validation';
+import { type FormError } from 'helpers/subscriptionsForms/validation';
 import { directDebitReducer as directDebit } from 'components/directDebit/directDebitReducer';
 import {
   marketingConsentReducerFor,
   type State as MarketingConsentState,
 } from 'components/marketingConsent/marketingConsentReducer';
-import { getSignoutUrl } from 'helpers/externalLinks';
 import { isTestUser } from 'helpers/user/user';
 import type { ErrorReason } from 'helpers/errorReasons';
 import { createUserReducer } from 'helpers/user/userReducer';
 import { fromCountry } from 'helpers/internationalisation/countryGroup';
 import type { ProductPrices } from 'helpers/productPrice/productPrices';
+import { validateForm } from 'pages/digital-subscription-checkout/helpers/validation';
 import type { Action } from './digitalSubscriptionCheckoutActions';
-import { setFormErrors } from './digitalSubscriptionCheckoutActions';
 import { getUser } from './helpers/user';
 import { showPaymentMethod, countrySupportsDirectDebit } from './helpers/paymentProviders';
 
@@ -103,58 +102,6 @@ function getEmail(state: State): string {
 }
 
 // ----- Functions ----- //
-
-function getErrors(fields: FormFields): FormError<FormField>[] {
-  return validate([
-    {
-      rule: nonEmptyString(fields.firstName),
-      error: formError('firstName', 'Please enter a value.'),
-    },
-    {
-      rule: nonEmptyString(fields.lastName),
-      error: formError('lastName', 'Please enter a value.'),
-    },
-    {
-      rule: nonEmptyString(fields.addressLine1),
-      error: formError('addressLine1', 'Please enter a value'),
-    },
-    {
-      rule: nonEmptyString(fields.townCity),
-      error: formError('townCity', 'Please enter a value'),
-    },
-    {
-      rule: nonEmptyString(fields.postcode),
-      error: formError('postcode', 'Please enter a value'),
-    },
-    {
-      rule: notNull(fields.country),
-      error: formError('country', 'Please select a country.'),
-    },
-    {
-      rule: fields.country === 'US' || fields.country === 'CA' ? notNull(fields.stateProvince) : true,
-      error: formError(
-        'stateProvince',
-        fields.country === 'CA' ? 'Please select a province/territory.' : 'Please select a state.',
-      ),
-    },
-  ]);
-}
-
-const formIsValid = () => (dispatch: Function, getState: () => State): boolean => {
-  const errors = getErrors(getFormFields(getState()));
-  return errors.length === 0;
-};
-
-const signOut = () => { window.location.href = getSignoutUrl(); };
-
-function validateForm(dispatch: Dispatch<Action>, state: State) {
-  const errors = getErrors(getFormFields(state));
-  const valid = errors.length === 0;
-  if (!valid) {
-    dispatch(setFormErrors(errors));
-  }
-  return valid;
-}
 
 function submitForm(dispatch: Dispatch<Action>, state: State) {
   if (validateForm(dispatch, state)) {
@@ -271,11 +218,7 @@ function initReducer(initialCountry: IsoCountry) {
 
 export {
   initReducer,
-  getErrors,
   getFormFields,
   getEmail,
-  signOut,
-  formIsValid,
   submitForm,
-  validateForm,
 };
